Тестовое задание для KK

+++ kkspb/src/main/java/ru/kk/repository/UrlRepository.java
@@ -3,7 +3,7 @@ package ru.kk.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.kk.model.UrlTable;
 
-public interface UrlRepository extends CrudRepository<UrlTable, Integer> {
+public interface UrlRepository extends CrudRepository<UrlTable, Long> {
     UrlTable findByShortUrl(String shortUrl);
     UrlTable findByUrl(String url);
 }
